refactor: move linkItem direc
